requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
alembic
celery
redis
pytest
requests
pyyaml
allure-pytest
pydantic
croniter
httpx
python-multipart
